--- a/package.json
+++ b/package.json
@@ -14,7 +14,7 @@
     "path-is-absolute": "^1.0.1",
     "process": "~0.11.0",
     "through2": "^2.0.0",
-    "undeclared-identifiers": "^1.0.0",
+    "undeclared-identifiers": "^1.1.2",
     "xtend": "^4.0.0"
   },
   "devDependencies": {