--- a/test/isbuffer.js
+++ b/test/isbuffer.js
@@ -4,6 +4,9 @@ var bpack = require('browser-pack');
 var insert = require('../');
 var concat = require('concat-stream');
 var vm = require('vm');
+// use is-buffer instead of builtin Buffer.isBuffer. The builtin
+// does `instanceof` which does not work on the browserified version
+var isBuffer = require('is-buffer');
 
 test('isbuffer', function (t) {
     t.plan(5);
@@ -20,6 +23,23 @@ test('isbuffer', function (t) {
     }));
     deps.write({ transform: inserter, global: true });
     deps.end({ id: 'main', file: __dirname + '/isbuffer/main.js' });
+});
+
+test('isbuffer (and buffer)', function (t) {
+    t.plan(5);
+    var deps = mdeps()
+    var pack = bpack({ raw: true, hasExports: true });
+    deps.pipe(pack).pipe(concat(function (src) {
+        var c = { global: {} };
+        vm.runInNewContext(src, c);
+        t.equal(c.require('main')(c.require('main').a()), true, 'is a buffer');
+        t.equal(c.require('main')('wow'), false, 'is not a buffer');
+        t.equal(isBuffer(c.require('main').a()), true, 'is a buffer');
+        t.ok(/require\("buffer"\)/.test(src), 'buffer required in source')
+        t.equal(c.require('main').a().toString('hex'), 'abcd', 'is a buffer');
+    }));
+    deps.write({ transform: inserter, global: true });
+    deps.end({ id: 'main', file: __dirname + '/isbuffer/both.js' });
 });
 
 function inserter (file) {