--- a/test/isbuffer/both.js
+++ b/test/isbuffer/both.js
@@ -1,0 +1,6 @@
+module.exports = function (buf) {
+    return Buffer.isBuffer(buf);
+};
+module.exports.a = function () {
+    return Buffer.from('abcd', 'hex');
+};