--- a/test/isbuffer/new.js
+++ b/test/isbuffer/new.js
@@ -1,0 +1,6 @@
+module.exports = function (buf) {
+    return Buffer.isBuffer(buf);
+};
+module.exports.a = function () {
+    return new Buffer('abcd');
+};