undeclared-identifiers collects references to undeclared identifiers and their properties. It does much less work than lexical-scope so it is quite a bit faster.

before:

    # jquery
    ok ~432 ms (0 s + 432086435 ns)

after:

    # jquery
    ok ~223 ms (0 s + 223387040 ns)