@goto-bus-stop 2018-06-15
ya. the consequences of a change like this are difficult to predict so hopefully it doesn't break anyone. if it does, i'll revert.